Randomise the order of experiments to run
Randomise the order of datasets and preprocessing methods to reduce risk
of encountering a race condition when running multiple experiments in
parallel. The race condition in this context means executing the same
run, i.e the same combination of a dataset and a preprocessing method,
at the same time from two different executions of experiment.